PR #717: end to end test with a handler
