+++ powertools-idempotency/src/test/java/software/amazon/lambda/powertools/idempotency/DynamoDBConfig.java
@@ -1,0 +1,80 @@
+package software.amazon.lambda.powertools.idempotency;
+
+import com.amazonaws.services.dynamodbv2.local.main.ServerRunner;
+import com.amazonaws.services.dynamodbv2.local.server.DynamoDBProxyServer;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.http.urlconnection.UrlConnectionHttpClient;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
+import software.amazon.awssdk.services.dynamodb.model.*;
+
+import java.io.IOException;
+import java.net.ServerSocket;
+import java.net.URI;
+
+public class DynamoDBConfig {
+    protected static final String TABLE_NAME = "idempotency_table";
+    protected static DynamoDBProxyServer dynamoProxy;
+    protected static DynamoDbClient client;
+
+    @BeforeAll
+    public static void setupDynamo() {
+        System.setProperty("sqlite4java.library.path", "src/test/native-libs");
+        int port = getFreePort();
+        try {
+            dynamoProxy = ServerRunner.createServerFromCommandLineArgs(new String[]{
+                    "-inMemory",
+                    "-port",
+                    Integer.toString(port)
+            });
+            dynamoProxy.start();
+        } catch (Exception e) {
+            throw new RuntimeException();
+        }
+
+        client = DynamoDbClient.builder()
+                .httpClient(UrlConnectionHttpClient.builder().build())
+                .region(Region.EU_WEST_1)
+                .endpointOverride(URI.create("http://localhost:" + port))
+                .credentialsProvider(StaticCredentialsProvider.create(
+                        AwsBasicCredentials.create("FAKE", "FAKE")))
+                .build();
+
+        client.createTable(CreateTableRequest.builder()
+                .tableName(TABLE_NAME)
+                .keySchema(KeySchemaElement.builder().keyType(KeyType.HASH).attributeName("id").build())
+                .attributeDefinitions(
+                        AttributeDefinition.builder().attributeName("id").attributeType(ScalarAttributeType.S).build()
+                )
+                .billingMode(BillingMode.PAY_PER_REQUEST)
+                .build());
+
+        DescribeTableResponse response = client.describeTable(DescribeTableRequest.builder().tableName(TABLE_NAME).build());
+        if (response == null) {
+            throw new RuntimeException("Table was not created within expected time");
+        }
+    }
+
+    @AfterAll
+    public static void teardownDynamo() {
+        try {
+            dynamoProxy.stop();
+        } catch (Exception e) {
+            throw new RuntimeException();
+        }
+    }
+
+    private static int getFreePort() {
+        try {
+            ServerSocket socket = new ServerSocket(0);
+            int port = socket.getLocalPort();
+            socket.close();
+            return port;
+        } catch (IOException ioe) {
+            throw new RuntimeException(ioe);
+        }
+    }
+}
